Set and retrieve mappings for facile objects.

aes_map retrieves all aesthetic maps, or the map for an individual aesthetic when aes_name is specified.

aes_map()<- resets the map, wholesale.

aes_map(x, aes_name = NULL, ...)

# S3 method for default
aes_map(x, aes_name = NULL, ...)

aes_map(x, aes_name = NULL) <- value

# S3 method for default
aes_map(x, aes_name = NULL) <- value

Arguments

x
an orbitrary object

aes_name
optional aes_name you want to individually set or retrieve, otherweise entire aesthetic map list is set/returned

Value

the getter returns the whole aesthetic map when is.null(aes_name), or the specific mapping for the named aesthetic. The setter sets the specific aes_name when specified, otherwise sets the aesthetics in the named value list

See also

Examples

x <- data.frame(a = 1:5, b = letters[1:5])
aes_map(x, "color") <- create_color_map(x$b)
aes_map(x, "size") <- "all sizes"
aes_map(x, "color")

#>         a         b         c         d         e 
#> "#999999" "#FF7F00" "#6495ED" "#E41A1C" "#4DAF4A" 

# simultaneously updates color map and adds shape map
aes_map(x) <- list(color = create_color_map(x$b, "Set3"),
                   shape = "all shapes")
aes_map(x)

#> $color
#>         a         b         c         d         e 
#> "#8DD3C7" "#FFFFB3" "#BEBADA" "#FB8072" "#80B1D3" 
#> 
#> $size
#> [1] "all sizes"
#> 
#> $shape
#> [1] "all shapes"
#>