Maps shapes to categorical values

Map unique leves of vals to different shapes. Only works for categorical variables.

create_shape_map(vals, map = NULL, ..., is3d = FALSE)

Arguments

vals
a vector of categorical values

map
a map definition. By default we use pch symbol identifiers.

Value

a named vector. names() are the unique values in vals, and values are the different shapes (pch integers)

Details

TODO: Use plotly shapes (currently we use base R pch). This webpage shows you the symbols and how to generate them: http://www.r-graph-gallery.com/125-the-plotlys-symbols/

Examples

# This isn't a real example. It is the code from the aforementioned page
# that generates the plotly shapes.
library(plotly)

#> Loading required package: ggplot2

#> 
#> Attaching package: ‘plotly’

#> The following object is masked from ‘package:ggplot2’:
#> 
#>     last_plot

#> The following object is masked from ‘package:stats’:
#> 
#>     filter

#> The following object is masked from ‘package:graphics’:
#> 
#>     layout

data=expand.grid(c(1:6) , c(1:6))
data=cbind(data , my_symbol=c(1:36))
data=data[data$my_symbol<33 , ]

# Make the graph
my_graph=plot_ly(data , x=~Var1 , y=~Var2 , type="scatter",
                 mode="markers+text" , hoverinfo="text", text=~my_symbol,
                 textposition = "bottom right",
                 marker=list(symbol=~my_symbol, size=40, color="red",
                             opacity=0.7)) %>%
  layout(
    hovermode="closest",
    yaxis=list(autorange="reversed", title="",
               tickfont=list(color="white")) ,
    xaxis=list( title="" , tickfont=list(color="white"))
  )
# show graph
my_graph

#> Warning: `arrange_()` is deprecated as of dplyr 0.7.0.
#> Please use `arrange()` instead.
#> See vignette('programming') for more help
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.