Drives faceting-like behavior for (internal) plotly-generating functions.

Other FacileViz functions call this function to drive faceting behavior. This function is largely meant to be internal and is purposefully not exported yet.

maybe_facet(
  plotfn,
  plotdat,
  facet_aes,
  nrows = NULL,
  has_legend = FALSE,
  plot_type = NULL,
  ...,
  height = 600,
  width = height + (height/6),
  facet_height = 300,
  facet_width = 370,
  legendside = NULL,
  showlegend = TRUE,
  widths = NULL,
  heights = NULL,
  margin = 0.02,
  shareX = TRUE,
  shareY = TRUE,
  titleX = shareX,
  titleY = shareY,
  which_layout = "merge"
)

Arguments

facet_aes
The name of a column in dat to use for faceting. If this is a length-2 vector, then we try to do a facet_grid (undone.)

nrows
number of rows for laying out plots in a grid-like structure. Only used if no domain is already specified.

...
One of the following

any number of plotly/ggplot2 objects.

a list of plotly/ggplot2 objects.

a tibble with one list-column of plotly/ggplot2 objects.

widths
relative width of each column on a 0-1 scale. By default all columns have an equal relative width.

heights
relative height of each row on a 0-1 scale. By default all rows have an equal relative height.

margin
either a single value or four values (all between 0 and 1). If four values are provided, the first is used as the left margin, the second is used as the right margin, the third is used as the top margin, and the fourth is used as the bottom margin. If a single value is provided, it will be used as all four margins.

shareX
should the x-axis be shared amongst the subplots?

shareY
should the y-axis be shared amongst the subplots?

titleX
should x-axis titles be retained?

titleY
should y-axis titles be retained?

which_layout
adopt the layout of which plot? If the default value of "merge" is used, layout options found later in the sequence of plots will override options found earlier in the sequence. This argument also accepts a numeric vector specifying which plots to consider when merging.