Plot expression of a gene from a DGEList (this function is not facile)

This is a conveniencde function that takes your expression object (DGEList, or "voomed" object) and plots the expression of one gene (gene) across the entire dataset grouped by the sample covariate specified by group. You can optionally color each point by a second sample covariate specified by the color_by parameter.

# y is a DGEList
plot_expression(y, gene = "Fxyd6", group = "genotype", color_by = "source")

plot_expression(
  y,
  gene,
  group = "group",
  color_by = NULL,
  wrap_by = "symbol",
  inorder = TRUE,
  ...
)

Arguments

y
the DGEList or voomed object your data is in

gene
the name of the gene identifier or symbol you want to plot.

group
what column in the $samples (or $targets, for vm) pheno table to use to plot split expression across the x axis

color_by
name of column (like group param) to color your points by

wrap_by
the column in the dataset to wrap the results by. If length(gene) > 1, you're going to need to wrwap.

Value

a ggplot object