Augment a facile object with a color aesthetic

This function augments a variety of object with color mappings. For now we assume objects are only data.frames or tbls.

with_color(x, aesthetic, aes_map = NULL, ..., .default_color = I("black"))

# S3 method for data.frame
with_color(
  x,
  aesthetic = NULL,
  aes_map = NULL,
  out_column = ".color_aes.",
  ...,
  .default_color = "black"
)

# S3 method for tbl
with_color(
  x,
  aesthetic = NULL,
  aes_map = NULL,
  out_column = ".color_aes.",
  ...,
  .default_color = I("black")
)

Arguments

x
an object to colorize (currently just data.frame or tbls)

aesthetic
an "attribute" from x to color. The types of object this argument can be is specific to the class of x. For instance, for a data.frame, aesthetic would be a character vector that specifices the column(s) used to map observations to colors.

aes_map
the "logic" to use to map aesthetic to color. This can be something like the name of an RColorBrewer map.

...
moar things

.default_color
the default color used when only one level exists

out_column
the name of the column that is appended to x that will hold the colors

Value

the object x with an upated aes_map(x, "color")

x with an additional out_column column that holds the colors for the observations in the rows of x

data.frames and tibbles

data.frames and tibbles are augmented with:

A .color_aes.variable column that hold the value of the element "level" for the color mapping

.color_aes.value column that holds the actual color for the observation

an updated aes_map(x, "color")` color map that has the color map

The reason we add a .color_aes.variable is because the user may decide to color observations by combinations of columns, not just a single column.

See also

Examples

x <- data.frame(
  category = sample(c("a", "b", "c"), 20, replace = TRUE),
  subcat = sample(c("y", "z"), 20, replace = TRUE),
  zscore = rnorm(20),
  score = runif(20, min = 0, max = 1))
x <- with_color(x, aesthetic = "category", "Set3")
y <- with_color(x, aesthetic = "category", "Set1")

#> Warning: .aes_varname column `.color_aes.variable`already exists in data.frame

#> Warning: .aes_valname column `.color_aes.value`already exists in data.frame

z <- with_color(x, aesthetic = "category", "viridis")

#> Warning: .aes_varname column `.color_aes.variable`already exists in data.frame

#> Warning: .aes_valname column `.color_aes.value`already exists in data.frame

m <- with_color(x, aesthetic = "score")

#> Warning: .aes_varname column `.color_aes.variable`already exists in data.frame

#> Warning: .aes_valname column `.color_aes.value`already exists in data.frame